Move Read, Write, Report to the package config and make them methods
