+++ examples/src/main/scala/zio/config/examples/CoproductExample.scala
@@ -32,7 +32,7 @@ object CoproductExample extends App {
   val source: ConfigSource =
     mapSource(validConfigForSampleConfig)
 
-  assert(runtime.unsafeRun(read(prodOrDev).run.provide(source))._2 == Left(Prod(Ldap("v1"), DbUrl("v2"))))
+  assert(runtime.unsafeRun(read(prodOrDev).provide(source))._2 == Left(Prod(Ldap("v1"), DbUrl("v2"))))
 
   val validConfigForAnotherConfig =
     Map(
@@ -45,7 +45,7 @@ object CoproductExample extends App {
   val anotherSource: ConfigSource =
     mapSource(validConfigForAnotherConfig)
 
-  assert(runtime.unsafeRun(read(prodOrDev).run.provide(anotherSource))._2 == Right(Dev("v3", 1, 2.0)))
+  assert(runtime.unsafeRun(read(prodOrDev).provide(anotherSource))._2 == Right(Dev("v3", 1, 2.0)))
 
   val invalidConfig =
     Map(
@@ -78,7 +78,7 @@ object CoproductExample extends App {
     )
 
   assert(
-    runtime.unsafeRun(read(prodOrDev).run.provide(mapSource(allConfigsExist)))._2 ==
+    runtime.unsafeRun(read(prodOrDev).provide(mapSource(allConfigsExist)))._2 ==
       Left(Prod(Ldap("v1"), DbUrl("v2")))
   )
 }
