+++ zio-config/src/test/scala/zio/config/CoproductTest.scala
@@ -69,7 +69,7 @@ object CoproductTest extends Properties("Coproduct support") with TestSupport {
   final case class EnterpriseAuth(ldap: Ldap, dburl: DbUrl)
   final case class PasswordAuth(user: String, count: Int, factor: Double)
 
-  private def readLeft(p: TestParams): IO[ReadErrors, Either[EnterpriseAuth, PasswordAuth]] = {
+  private def readLeft(p: TestParams): IO[List[ReadError], Either[EnterpriseAuth, PasswordAuth]] = {
     val enterprise: Config[EnterpriseAuth] =
       (string(p.kLdap).xmap(Ldap)(_.value) <*> string(p.kDbUrl).xmap(DbUrl)(_.value))(
         EnterpriseAuth.apply,
@@ -83,7 +83,7 @@ object CoproductTest extends Properties("Coproduct support") with TestSupport {
     val authConfig: Config[Either[EnterpriseAuth, PasswordAuth]] =
       enterprise or password
 
-    read(authConfig).run
+    read(authConfig)
       .map(_._2)
       .provide(
         mapSource(
@@ -95,7 +95,7 @@ object CoproductTest extends Properties("Coproduct support") with TestSupport {
       )
   }
 
-  private def readRight(p: TestParams): IO[ReadErrors, Either[EnterpriseAuth, PasswordAuth]] = {
+  private def readRight(p: TestParams): IO[List[ReadError], Either[EnterpriseAuth, PasswordAuth]] = {
     val enterprise: Config[EnterpriseAuth] =
       (string(p.kLdap).xmap(Ldap)(_.value) <*> string(p.kDbUrl).xmap(DbUrl)(_.value))(
         EnterpriseAuth.apply,
@@ -109,7 +109,7 @@ object CoproductTest extends Properties("Coproduct support") with TestSupport {
     val authConfig: Config[Either[EnterpriseAuth, PasswordAuth]] =
       enterprise or password
 
-    read(authConfig).run
+    read(authConfig)
       .map(_._2)
       .provide(
         mapSource(
@@ -124,7 +124,7 @@ object CoproductTest extends Properties("Coproduct support") with TestSupport {
 
   private def readWithErrors(
     p: TestParams
-  ): ZIO[Any, Nothing, Either[ReadErrors, Either[EnterpriseAuth, PasswordAuth]]] = {
+  ): ZIO[Any, Nothing, Either[List[ReadError], Either[EnterpriseAuth, PasswordAuth]]] = {
     val enterprise: Config[EnterpriseAuth] =
       (string(p.kLdap).xmap(Ldap)(_.value) <*> string(p.kDbUrl).xmap(DbUrl)(_.value))(
         EnterpriseAuth.apply,
@@ -138,7 +138,7 @@ object CoproductTest extends Properties("Coproduct support") with TestSupport {
     val authConfig: Config[Either[EnterpriseAuth, PasswordAuth]] =
       enterprise or password
 
-    read(authConfig).run
+    read(authConfig)
       .map(_._2)
       .provide(
         mapSource(
@@ -153,7 +153,7 @@ object CoproductTest extends Properties("Coproduct support") with TestSupport {
       .either
   }
 
-  private def readChooseLeftFromBoth(p: TestParams): IO[ReadErrors, Either[EnterpriseAuth, PasswordAuth]] = {
+  private def readChooseLeftFromBoth(p: TestParams): IO[List[ReadError], Either[EnterpriseAuth, PasswordAuth]] = {
     val enterprise: Config[EnterpriseAuth] =
       (string(p.kLdap).xmap(Ldap)(_.value) <*> string(p.kDbUrl).xmap(DbUrl)(_.value))(
         EnterpriseAuth.apply,
@@ -167,7 +167,7 @@ object CoproductTest extends Properties("Coproduct support") with TestSupport {
     val authConfig: Config[Either[EnterpriseAuth, PasswordAuth]] =
       enterprise or password
 
-    read(authConfig).run
+    read(authConfig)
       .map(_._2)
       .provide(
         mapSource(
